requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
pydantic-settings==2.0.3  # Required for BaseSettings in Pydantic v2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.0
email-validator==2.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # PostgreSQL adapter